Moving iree/base/math.h to iree/base/internal/.
