+++ iree/hal/vulkan/CMakeLists.txt
@@ -85,6 +85,7 @@ iree_cc_library(
     iree::base::api
     iree::base::core_headers
     iree::base::flatcc
+    iree::base::internal
     iree::base::logging
     iree::base::status
     iree::base::synchronization
