+++ iree/hal/vulkan/vulkan_device.cc
@@ -20,7 +20,7 @@
 #include "absl/container/inlined_vector.h"
 #include "absl/memory/memory.h"
 #include "absl/strings/str_cat.h"
-#include "iree/base/math.h"
+#include "iree/base/internal/math.h"
 #include "iree/base/status.h"
 #include "iree/base/tracing.h"
 #include "iree/hal/vulkan/api.h"
