+++ iree/task/affinity_set.h
@@ -16,7 +16,7 @@
 #define IREE_TASK_AFFINITY_SET_H_
 
 #include "iree/base/internal/atomics.h"
-#include "iree/base/math.h"
+#include "iree/base/internal/math.h"
 #include "iree/task/tuning.h"
 
 #ifdef __cplusplus
