+++ iree/task/executor_impl.h
@@ -15,8 +15,8 @@
 #ifndef IREE_TASK_EXECUTOR_IMPL_H_
 #define IREE_TASK_EXECUTOR_IMPL_H_
 
+#include "iree/base/internal/math.h"
 #include "iree/base/internal/prng.h"
-#include "iree/base/math.h"
 #include "iree/base/synchronization.h"
 #include "iree/base/tracing.h"
 #include "iree/task/affinity_set.h"
