+++ iree/task/pool.c
@@ -15,7 +15,7 @@
 #include "iree/task/pool.h"
 
 #include "iree/base/internal/debugging.h"
-#include "iree/base/math.h"
+#include "iree/base/internal/math.h"
 
 // Minimum byte size of a block in bytes, including the tasks as well as the
 // allocation header. This is here to allow us to reduce the number of times
