+++ iree/task/post_batch.c
@@ -14,7 +14,7 @@
 
 #include "iree/task/post_batch.h"
 
-#include "iree/base/math.h"
+#include "iree/base/internal/math.h"
 #include "iree/base/synchronization.h"
 #include "iree/base/threading.h"
 #include "iree/base/tracing.h"
