+++ iree/task/topology.c
@@ -19,7 +19,7 @@
 #include <stdio.h>
 
 #include "iree/base/internal/debugging.h"
-#include "iree/base/math.h"
+#include "iree/base/internal/math.h"
 #include "iree/base/tracing.h"
 #include "iree/task/tuning.h"
