+++ iree/vm/CMakeLists.txt
@@ -172,6 +172,7 @@ iree_cc_library(
     iree::base::api
     iree::base::core_headers
     iree::base::flatcc
+    iree::base::internal
     iree::base::tracing
     iree::schemas::bytecode_module_def_c_fbs
   PUBLIC
